Add files via upload
The ERD shows the relationships between the different tables in the database. For example, the Games table is related to the City table through the Games_City table, which stores the ID of the game and the ID of the city where it was held. Similarly, the Competitor_Event table stores the ID of the event, the ID of the athlete who competed in it, and the ID of the medal won. The ERD helps in visualizing the database structure and its relationships.